Add Navbar and Toaster components
